refactor: package 별로 정리
