+++ src/main/java/Application.java
@@ -1,4 +1,4 @@
-package racingcar;
+import domain.CarRacing;
 
 public class Application {
     public static void main(String[] args) {
