+++ src/main/java/constants/SystemMessage.java
@@ -1,0 +1,4 @@
+package constants;
+
+public class SystemMessage {
+}
